Split the examples directory into multiple separate examples
